AppState 初期化エラー修正と証明書パス判定の統一 (#38)
* Fix AppState fields and cert path check

* refactor: split backend utilities

* fix: harden extraction and ids

* chore: bump version to 1.2.0 and remove File Version comments

- Update SYSTEM_VERSION to 1.2.0 across all files (major refactoring)
- Remove File Version comments from all files (git tracks history)
- Update CHANGELOG with refactoring details
- Update documentation (README, AGENTS, CLAUDE)


---------

  refactor: modular backend structure (v1.2.0)

  - Split web_server.py into app.py + backend/ modules
  - Remove File Version comments (use git history)
  - Harden PDF extraction and file scanning
  - Expand file_id to 12 chars
