+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 # 変更履歴
 
-System Version: 1.1.11
+System Version: 1.2.0
 
-## 1.1.11
+## 1.2.0
 
-- 2026-01-15: 改善: web_server.py を廃止し、app.py + backend/ 構成へ整理。起動方法を更新。
+- 2026-01-15: **リファクタリング**: web_server.py (4147行) を廃止し、モジュール分割された backend/ 構成へ移行。
+  - `app.py`: エントリポイント
+  - `backend/routes.py`: APIルートとアプリケーション状態
+  - `backend/search.py`: 検索ロジック
+  - `backend/index_ops.py`: インデックス操作
+  - `backend/extractors.py`: ファイル読み込み・テキスト抽出
+  - `backend/text_utils.py`: テキスト正規化
+  - `backend/utils.py`: ユーティリティ関数
+  - `backend/config.py`: 設定定数
+- 2026-01-15: 整理: File Version コメントを全ファイルから削除（git で履歴管理するため不要）。
+- 2026-01-15: 修正: PDFフォールバック抽出でページごとにファイル位置をリセットし、欠落リスクを低減。
+- 2026-01-15: 改善: 抽出失敗時のログ出力を追加し、原因特定を容易化。
+- 2026-01-15: 改善: file_id を12桁に拡張し、/api/detail のバリデーションを更新。
+- 2026-01-15: 改善: process 非共有時の build_memory_pages を routes と整合させ、表示・正規化の差異を解消。
+- 2026-01-15: 改善: scan_files の権限/OSエラー耐性を追加し、部分的に取得可能に。
+- 2026-01-15: 修正: AppState の未定義フィールドが初期化エラーになる問題を修正。
+- 2026-01-15: 修正: HTTPS 判定の証明書パス解決を BASE_DIR 基準に統一。
 - 2026-01-14: 改善: インデックス構築時に対象フォルダが見つからない場合は warn ログを出力。
 
 ## 1.1.10
