+++ README.md
@@ -1,7 +1,6 @@
 # Full-Search-PDFs
 
-System Version: 1.1.11
-README File Version: 1.1.12
+System Version: 1.2.0
 
 ローカル/社内フォルダ内の PDF/Office 文書を全文検索する FastAPI アプリです。
 UI は `static/` 配下で提供されます。
@@ -188,6 +187,5 @@ REBUILD_SCHEDULE="03:00"
 - `SEARCH_PROCESS_SHARED=1`（既定）ではページ本文が共有 mmap になり増加は抑制されます
 - PDFのレイアウトによっては抽出順序が混ざる場合があります
 
-## 変更履歴とバージョン運用
+## 変更履歴
 - 変更ごとに `CHANGELOG.md` を日本語で更新してください。
-- 各ファイルには System Version と File Version を記載しています。
