+++ backend/__init__.py
@@ -1,4 +1,3 @@
-SYSTEM_VERSION = "1.1.11"
-# File Version: 1.0.0
+SYSTEM_VERSION = "1.2.0"
 
 from .routes import create_app  # noqa: F401
