+++ backend/routes.py
@@ -2,16 +2,14 @@ import asyncio
 import csv
 import gzip
 import hashlib
+import io
 import json
 import mmap
 import os
 import pickle
 import re
-import socket
-import subprocess
 import threading
 import time
-import unicodedata
 from collections import defaultdict, OrderedDict
 from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
 from contextlib import asynccontextmanager
@@ -19,14 +17,12 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Tuple
 
-SYSTEM_VERSION = "1.1.11"
-# File Version: 1.0.0
+SYSTEM_VERSION = "1.2.0"
 from fastapi import FastAPI, HTTPException, Query
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import FileResponse, JSONResponse, StreamingResponse
 from fastapi.staticfiles import StaticFiles
 from pydantic import BaseModel, Field, field_validator
-import io
 
 from .config import (
     ALLOWED_EXTS,
@@ -54,874 +50,54 @@ from .config import (
     load_config,
 )
 from .context import AppContext, AppState
-try:
-    from colorama import Fore, Style, init as colorama_init
-except Exception:
-    Fore = None
-    Style = None
-    colorama_init = None
-
-# --- 外部依存ロジック ---
-from docx import Document
-from openpyxl import load_workbook
-from pdfminer.high_level import extract_pages, extract_text as miner_extract_text
-from pdfminer.layout import LAParams, LTTextBoxHorizontal, LTTextBoxVertical, LTTextContainer
-from pdfminer.pdfpage import PDFPage
-
-# --- ユーティリティ ---
-CID_PATTERN = re.compile(r"\(cid:\d+\)", re.IGNORECASE)
-INVISIBLE_SEPARATORS_PATTERN = re.compile(r"[\u200b\u200c\u200d\u2060\ufeff\u00a0\u202f]")
-JP_CHAR_CLASS = r"\u3040-\u30FF\u3400-\u4DBF\u4E00-\u9FFF\uF900-\uFAFF\uFF66-\uFF9F"
-OCR_PAGE_PATTERN = re.compile(r"^\s*-{2,}\s*Page\s*(\d+)?\s*-{2,}\s*$", re.IGNORECASE)
-
-
-def strip_cid(text: str) -> str:
-    if not text:
-        return ""
-    return CID_PATTERN.sub("", text)
-
-
-def normalize_invisible_separators(text: str) -> str:
-    if not text:
-        return ""
-    return INVISIBLE_SEPARATORS_PATTERN.sub(" ", text)
-
-
-def build_flexible_keyword_regex(keyword: str) -> re.Pattern | None:
-    compact = re.sub(r"\s+", "", keyword or "")
-    if len(compact) < 2 or len(compact) > 64:
-        return None
-    parts = [re.escape(ch) for ch in compact]
-    pattern = r"\s*".join(parts)
-    return re.compile(pattern, re.IGNORECASE)
-
-
-def apply_space_mode(text: str, mode: str) -> str:
-    if not text:
-        return ""
-    if mode == "all":
-        return re.sub(r"\s+", "", text)
-    if mode == "jp":
-        pattern = rf"(?<=[{JP_CHAR_CLASS}])\s+(?=[{JP_CHAR_CLASS}])"
-        return re.sub(pattern, "", text)
-    return text
-
-
-def normalize_text(text: str) -> str:
-    """キーワード検索のための正規化."""
-    if not text:
-        return ""
-    text = strip_cid(text)
-    text = unicodedata.normalize("NFKC", text)
-    text = normalize_invisible_separators(text)
-    text = text.lower()
-    text = re.sub(r"[\n\t\r]", " ", text)
-    text = re.sub(r"\s+", " ", text)
-    return text
-
-
-def normalize_text_minimal(text: str) -> str:
-    """改行/不可視の最小整形のみ行う."""
-    if not text:
-        return ""
-    text = strip_cid(text)
-    text = normalize_invisible_separators(text)
-    text = re.sub(r"[\n\t\r]", " ", text)
-    text = re.sub(r"\s+", " ", text)
-    return text.strip()
-
-
-def normalize_text_nfkc_casefold(text: str, compress_spaces: bool = True) -> str:
-    """NFKC + casefold の正規化."""
-    if not text:
-        return ""
-    text = strip_cid(text)
-    text = unicodedata.normalize("NFKC", text)
-    text = normalize_invisible_separators(text)
-    text = text.casefold()
-    text = re.sub(r"[\n\t\r]", " ", text)
-    if compress_spaces:
-        text = re.sub(r"\s+", " ", text)
-    return text
-
-
-def encode_norm_text(text: str) -> bytes:
-    if not text:
-        return b""
-    return text.encode("utf-8")
-
-
-def normalize_snippet_text(text: str) -> str:
-    text = strip_cid(text)
-    text = normalize_invisible_separators(text)
-    text = text.replace("\r", "")
-    text = re.sub(r"\s+", " ", text)
-    return text.strip()
-
-
-def normalize_detail_text(text: str) -> str:
-    text = strip_cid(text)
-    text = normalize_invisible_separators(text)
-    text = text.replace("\r", "")
-    text = re.sub(r"\n(?!　)", " ", text)
-    text = re.sub(r"[ \t]+", " ", text).strip()
-    return text
-
-
-def build_detail_text(raw_text: str, raw_hit_pos: int) -> str:
-    detail_pos = raw_hit_pos if raw_hit_pos != -1 else 0
-    detail_start = max(0, detail_pos - DETAIL_CONTEXT_PREFIX)
-    detail_end = min(len(raw_text), detail_start + DETAIL_WINDOW_SIZE)
-    detail_text_raw = raw_text[detail_start:detail_end]
-    return normalize_detail_text(detail_text_raw)
-
-
-def get_ipv4_addresses() -> List[str]:
-    ips: set[str] = set()
-    try:
-        hostname = socket.gethostname()
-        for item in socket.getaddrinfo(hostname, None, socket.AF_INET):
-            ip = item[4][0]
-            if ip and not ip.startswith("127."):
-                ips.add(ip)
-        for ip in socket.gethostbyname_ex(hostname)[2]:
-            if ip and not ip.startswith("127."):
-                ips.add(ip)
-    except Exception:
-        pass
-
-    if os.name == "nt":
-        try:
-            output = subprocess.check_output(["ipconfig"], text=True, errors="ignore")
-            for line in output.splitlines():
-                line = line.strip()
-                if "IPv4 Address" in line or "IPv4 アドレス" in line:
-                    parts = line.split(":")
-                    if len(parts) >= 2:
-                        ip = parts[-1].strip()
-                        if ip and not ip.startswith("127."):
-                            ips.add(ip)
-        except Exception:
-            pass
-    else:
-        try:
-            output = subprocess.check_output(["/sbin/ip", "-4", "addr"], text=True, errors="ignore")
-            for match in re.finditer(r"\\binet\\s+(\\d+\\.\\d+\\.\\d+\\.\\d+)", output):
-                ip = match.group(1)
-                if ip and not ip.startswith("127."):
-                    ips.add(ip)
-        except Exception:
-            pass
-    return sorted(ips)
-
-
-def split_by_ocr_markers(text: str) -> Dict[int, str]:
-    lines = text.split("\n")
-    pages: Dict[int, str] = {}
-    buffer: List[str] = []
-    current_page = 1
-    found_marker = False
-    for line in lines:
-        match = OCR_PAGE_PATTERN.match(line)
-        if match:
-            found_marker = True
-            if buffer:
-                pages[current_page] = "\n".join(buffer).strip()
-                buffer = []
-            page_hint = match.group(1)
-            if page_hint and page_hint.isdigit():
-                current_page = int(page_hint)
-            else:
-                current_page += 1
-            continue
-        buffer.append(line)
-    if buffer:
-        pages[current_page] = "\n".join(buffer).strip()
-    if not found_marker:
-        return {}
-    return {k: v for k, v in pages.items() if v}
-
-
-def find_break_pos(text: str, start: int, target: int, min_size: int, max_size: int) -> int:
-    end_limit = min(len(text), start + max_size)
-    target_pos = min(len(text), start + target)
-    window_start = min(len(text), start + min_size)
-    if window_start >= end_limit:
-        return end_limit
-
-    primary = set(["\n", " ", "\t"])
-    secondary = set("。．、，.!?？！)]}」』】）〕〉》】○×△□◇■-―—")
-
-    def scan_for(chars: set) -> int | None:
-        for pos in range(target_pos, end_limit):
-            if text[pos] in chars:
-                return pos
-        for pos in range(target_pos - 1, window_start - 1, -1):
-            if text[pos] in chars:
-                return pos
-        return None
-
-    hit = scan_for(primary)
-    if hit is None:
-        hit = scan_for(secondary)
-    if hit is None:
-        return min(start + target, len(text))
-    while hit + 1 < len(text) and text[hit + 1] in primary:
-        hit += 1
-    return hit + 1
-
-
-def split_text_by_size(
-    text: str,
-    target: int = 1200,
-    min_size: int = 600,
-    max_size: int = 1800,
-    overlap: int = 200,
-) -> Dict[int, str]:
-    pages: Dict[int, str] = {}
-    pos = 0
-    page_num = 1
-    length = len(text)
-    if length <= max_size:
-        cleaned = text.strip()
-        return {1: cleaned} if cleaned else {}
-    while pos < length:
-        remaining = length - pos
-        if remaining <= max_size:
-            chunk = text[pos:].strip()
-            if chunk:
-                pages[page_num] = chunk
-            break
-        split_pos = find_break_pos(text, pos, target, min_size, max_size)
-        if split_pos <= pos:
-            split_pos = min(pos + target, length)
-        if split_pos <= pos:
-            break
-        chunk = text[pos:split_pos].strip()
-        if chunk:
-            pages[page_num] = chunk
-            page_num += 1
-        if overlap > 0:
-            pos = max(split_pos - overlap, pos + 1)
-        else:
-            pos = split_pos
-    return pages
-
-
-def split_non_pdf_text(text: str) -> Dict[int, str]:
-    normalized = text.replace("\r\n", "\n").replace("\r", "\n")
-    pages = split_by_ocr_markers(normalized)
-    if pages:
-        return pages
-    return split_text_by_size(normalized)
-
-
-def read_text_file_safe(path: str) -> str:
-    encodings = ["utf-8", "cp932", "shift_jis", "euc-jp"]
-    for enc in encodings:
-        try:
-            with open(path, "r", encoding=enc) as f:
-                return f.read()
-        except Exception:
-            continue
-    return ""
-
-
-def read_csv_file_safe(path: str) -> str:
-    encodings = ["utf-8", "cp932", "shift_jis"]
-    for enc in encodings:
-        try:
-            with open(path, "r", encoding=enc, newline="") as f:
-                reader = csv.reader(f)
-                return "\n".join([" ".join(row) for row in reader])
-        except Exception:
-            continue
-    return ""
-
-
-def read_excel_rows_xlsx(path: str) -> List[Tuple[str, int, str]]:
-    try:
-        wb = load_workbook(path, read_only=True, data_only=True)
-    except Exception:
-        return []
-
-    rows: List[Tuple[str, int, str]] = []
-    try:
-        for ws in wb.worksheets:
-            sheet_name = ws.title or "Sheet"
-            row_idx = 0
-            for row in ws.iter_rows(values_only=True):
-                row_idx += 1
-                cells = [
-                    str(cell).strip()
-                    for cell in row
-                    if cell is not None and str(cell).strip()
-                ]
-                if cells:
-                    rows.append((sheet_name, row_idx, " ".join(cells)))
-    except Exception:
-        return []
-    finally:
-        try:
-            wb.close()
-        except Exception:
-            pass
-    return rows
-
-
-def read_excel_rows_xls(path: str) -> List[Tuple[str, int, str]]:
-    try:
-        import xlrd
-    except Exception:
-        return []
-    try:
-        wb = xlrd.open_workbook(path)
-    except Exception:
-        return []
-
-    rows: List[Tuple[str, int, str]] = []
-    try:
-        for sheet in wb.sheets():
-            sheet_name = sheet.name or "Sheet"
-            for r in range(sheet.nrows):
-                row_values = [
-                    str(sheet.cell_value(r, c)).strip()
-                    for c in range(sheet.ncols)
-                    if str(sheet.cell_value(r, c)).strip()
-                ]
-                if row_values:
-                    rows.append((sheet_name, r + 1, " ".join(row_values)))
-    except Exception:
-        return []
-    return rows
-
-
-def split_excel_rows(rows: List[Tuple[str, int, str]], rows_per_page: int = 40) -> Dict[str, str]:
-    pages: Dict[str, str] = {}
-    by_sheet: Dict[str, List[Tuple[int, str]]] = {}
-    for sheet, row_idx, text in rows:
-        by_sheet.setdefault(sheet, []).append((row_idx, text))
-
-    for sheet, sheet_rows in by_sheet.items():
-        sheet_rows.sort(key=lambda x: x[0])
-        for i in range(0, len(sheet_rows), rows_per_page):
-            chunk = sheet_rows[i : i + rows_per_page]
-            start_row = chunk[0][0]
-            end_row = chunk[-1][0]
-            label = f"{sheet} 行 {start_row}-{end_row}"
-            pages[label] = "\n".join(text for _, text in chunk).strip()
-    return pages
-
-
-def iter_text_containers(obj):
-    if isinstance(obj, LTTextContainer):
-        yield obj
-        return
-    for child in getattr(obj, "_objs", []):
-        yield from iter_text_containers(child)
-
-
-def order_text_boxes(boxes: List[Dict], page_bbox: Tuple[float, float, float, float]) -> str:
-    if not boxes:
-        return ""
-    x0, y0, x1, y1 = page_bbox
-    page_mid = (x0 + x1) / 2.0
-    vertical_count = sum(1 for b in boxes if b["vertical"])
-    horizontal_count = len(boxes) - vertical_count
-    use_vertical = vertical_count > horizontal_count and vertical_count >= 3
-
-    if use_vertical:
-        ordered = sorted(boxes, key=lambda b: (-b["x0"], -b["y1"]))
-    else:
-        left = [b for b in boxes if b["cx"] < page_mid]
-        right = [b for b in boxes if b["cx"] >= page_mid]
-        is_two_column = (
-            len(boxes) >= 6
-            and len(left) >= 2
-            and len(right) >= 2
-            and 0.3 < (len(left) / len(boxes)) < 0.7
-        )
-        if is_two_column:
-            left.sort(key=lambda b: (-b["y1"], b["x0"]))
-            right.sort(key=lambda b: (-b["y1"], b["x0"]))
-            ordered = left + right
-        else:
-            ordered = sorted(boxes, key=lambda b: (-b["y1"], b["x0"]))
-
-    return "\n".join(b["text"].strip() for b in ordered if b["text"].strip())
-
-
-def extract_pdf_text_by_layout(file_path: str) -> Dict[int, str]:
-    page_texts: Dict[int, str] = {}
-    laparams = LAParams()
-    for page_num, layout in enumerate(extract_pages(file_path, laparams=laparams), start=1):
-        boxes = []
-        for container in iter_text_containers(layout):
-            text = container.get_text()
-            if not text or not text.strip():
-                continue
-            x0, y0, x1, y1 = container.bbox
-            boxes.append(
-                {
-                    "text": text,
-                    "x0": x0,
-                    "y0": y0,
-                    "x1": x1,
-                    "y1": y1,
-                    "cx": (x0 + x1) / 2.0,
-                    "vertical": isinstance(container, LTTextBoxVertical),
-                }
-            )
-        if boxes:
-            ordered_text = order_text_boxes(boxes, layout.bbox)
-            if ordered_text:
-                page_texts[page_num] = ordered_text
-    return page_texts
-
-
-def extract_text_from_file(file_path: str) -> Dict[int, str]:
-    """各ファイルからテキストを抽出（PDFは高精度モード固定）."""
-    texts, _ = extract_text_from_file_with_reason(file_path)
-    return texts
-
-
-def empty_reason(file_path: str, fallback: str) -> str:
-    try:
-        if os.path.getsize(file_path) == 0:
-            return "空ファイル"
-    except Exception:
-        return fallback
-    return fallback
-
-
-def extract_text_from_file_with_reason(file_path: str) -> Tuple[Dict[int, str], str]:
-    """抽出結果と理由を返す."""
-    path_obj = Path(file_path)
-    ext = path_obj.suffix.lower()
-    page_texts: Dict[int, str] = {}
-    file_path_str = str(file_path)
-    reason = ""
-
-    try:
-        if ext == ".pdf":
-            try:
-                page_texts = extract_pdf_text_by_layout(file_path_str)
-            except Exception:
-                page_texts = {}
-            if not page_texts:
-                laparams = LAParams()
-                with open(file_path_str, "rb") as fp:
-                    pages = list(PDFPage.get_pages(fp, check_extractable=False))
-                    total_pages = len(pages)
-
-                with open(file_path_str, "rb") as fp:
-                    for i in range(total_pages):
-                        try:
-                            text = miner_extract_text(fp, page_numbers=[i], laparams=laparams)
-                            if text and text.strip():
-                                page_texts[i + 1] = text
-                        except Exception:
-                            continue
-            if not page_texts:
-                reason = empty_reason(file_path_str, "PDF抽出失敗")
-
-        elif ext == ".docx":
-            doc = Document(file_path_str)
-            full_text = []
-            for para in doc.paragraphs:
-                full_text.append(para.text)
-            for table in doc.tables:
-                for row in table.rows:
-                    for cell in row.cells:
-                        full_text.append(cell.text)
-            content = "\n".join(full_text)
-            if content:
-                page_texts = split_non_pdf_text(content)
-            else:
-                reason = empty_reason(file_path_str, "内容が空")
-
-        elif ext == ".txt":
-            content = read_text_file_safe(file_path_str)
-            if content:
-                page_texts = split_non_pdf_text(content)
-            else:
-                reason = empty_reason(file_path_str, "読み取り失敗または内容が空")
-
-        elif ext == ".csv":
-            content = read_csv_file_safe(file_path_str)
-            if content:
-                page_texts = split_non_pdf_text(content)
-            else:
-                reason = empty_reason(file_path_str, "読み取り失敗または内容が空")
-        elif ext == ".xlsx":
-            rows = read_excel_rows_xlsx(file_path_str)
-            if rows:
-                page_texts = split_excel_rows(rows)
-            else:
-                reason = empty_reason(file_path_str, "読み取り失敗または内容が空")
-        elif ext == ".xls":
-            rows = read_excel_rows_xls(file_path_str)
-            if rows:
-                page_texts = split_excel_rows(rows)
-            else:
-                reason = empty_reason(file_path_str, "読み取り失敗または内容が空")
-    except Exception as exc:
-        return {}, f"例外:{type(exc).__name__}"
-
-    if page_texts:
-        return page_texts, ""
-    return {}, reason or "抽出結果が空"
-
-
-def _find_raw_hit_position(
-    raw_for_positions: str,
-    raw_keywords: List[str],
-    space_mode: str,
-) -> int:
-    """生テキスト内でキーワードの位置を検索する共通ヘルパー."""
-    raw_hit_pos = -1
-    raw_lower = raw_for_positions.lower()
-    for kw in raw_keywords:
-        if not kw:
-            continue
-        raw_idx = raw_lower.find(kw.lower())
-        if raw_idx != -1:
-            raw_hit_pos = raw_idx
-            break
-
-    if raw_hit_pos == -1 and space_mode != "none":
-        for kw in raw_keywords:
-            if not kw:
-                continue
-            flex = build_flexible_keyword_regex(kw)
-            if not flex:
-                continue
-            match = flex.search(raw_for_positions)
-            if match:
-                raw_hit_pos = match.start()
-                break
-
-    return raw_hit_pos
-
-
-def _build_search_result(
-    entry: Dict,
-    raw_text: str,
-    raw_for_positions: str,
-    raw_hit_pos: int,
-    include_detail: bool,
-) -> Dict:
-    """検索結果のスニペットと詳細を生成する共通ヘルパー."""
-    snippet_start = max(0, raw_hit_pos - SNIPPET_PREFIX_CHARS) if raw_hit_pos != -1 else 0
-    snippet_end = min(len(raw_for_positions), snippet_start + SNIPPET_TOTAL_LENGTH)
-    snippet_raw = raw_for_positions[snippet_start:snippet_end]
-    snippet = f"...{normalize_snippet_text(snippet_raw)}..."
-
-    file_id = entry.get("fileId") or file_id_from_path(entry["path"])
-    page_raw = entry.get("pageRaw", entry.get("page"))
-    hit_pos = raw_hit_pos if raw_hit_pos >= 0 else 0
-    result = {
-        "file": entry["file"],
-        "path": entry["path"],
-        "displayPath": entry.get(
-            "displayPath", display_path_for_path(entry["path"], host_aliases)
-        ),
-        "page": entry["page"],
-        "context": snippet,
-        "detail_key": {
-            "file_id": file_id,
-            "page": page_raw,
-            "hit_pos": hit_pos,
-        },
-        "folderId": entry["folderId"],
-        "folderName": entry["folderName"],
-    }
-    if include_detail:
-        result["detail"] = build_detail_text(raw_text, hit_pos)
-    return result
-
-
-def search_text_logic(
-    entry: Dict,
-    norm_keyword_groups: List[List[str]],
-    raw_keywords: List[str],
-    search_mode: str,
-    range_limit: int,
-    space_mode: str,
-    normalize_mode: str,
-    include_detail: bool,
-) -> List[Dict]:
-    if not norm_keyword_groups:
-        return []
-
-    raw_text = entry["raw"]
-    raw_for_positions = normalize_invisible_separators(raw_text)
-    if normalize_mode == "normalized":
-        if space_mode == "all":
-            norm_text = entry.get("norm_cf_all") or entry["norm_all"]
-        elif space_mode == "jp":
-            norm_text = entry.get("norm_cf_jp") or entry["norm_jp"]
-        else:
-            norm_text = entry.get("norm_cf") or entry["norm"]
-    else:
-        if space_mode == "all":
-            norm_text = entry.get("norm_strict_all") or entry["norm_all"]
-        elif space_mode == "jp":
-            norm_text = entry.get("norm_strict_jp") or entry["norm_jp"]
-        else:
-            norm_text = entry.get("norm_strict") or entry["norm"]
-    is_match = False
-    match_pos = -1
-
-    if search_mode == "OR":
-        for group in norm_keyword_groups:
-            for k in group:
-                if not k:
-                    continue
-                idx = norm_text.find(k)
-                if idx != -1:
-                    is_match = True
-                    match_pos = idx
-                    break
-            if is_match:
-                break
-    elif search_mode == "AND":
-        for group in norm_keyword_groups:
-            if not any(k in norm_text for k in group if k):
-                return []
-        if range_limit == 0 or len(norm_keyword_groups) == 1:
-            first_group = norm_keyword_groups[0]
-            first_kw = next((k for k in first_group if k), "")
-            is_match = True
-            match_pos = norm_text.find(first_kw) if first_kw else 0
-        else:
-            positions = []
-            for group_id, group in enumerate(norm_keyword_groups):
-                for kw in group:
-                    if not kw:
-                        continue
-                    start = 0
-                    while True:
-                        idx = norm_text.find(kw, start)
-                        if idx == -1:
-                            break
-                        positions.append((idx, group_id))
-                        start = idx + 1
-            positions.sort()
-            left = 0
-            count_map = defaultdict(int)
-            unique_count = 0
-            target_unique = len(norm_keyword_groups)
-            min_window_len = float("inf")
-            best_start_pos = -1
-            for right in range(len(positions)):
-                r_idx, r_group_id = positions[right]
-                if count_map[r_group_id] == 0:
-                    unique_count += 1
-                count_map[r_group_id] += 1
-                while unique_count == target_unique:
-                    l_idx, l_group_id = positions[left]
-                    current_len = r_idx - l_idx
-                    if current_len <= range_limit:
-                        is_match = True
-                        if current_len < min_window_len:
-                            min_window_len = current_len
-                            best_start_pos = l_idx
-                    count_map[l_group_id] -= 1
-                    if count_map[l_group_id] == 0:
-                        unique_count -= 1
-                    left += 1
-            if is_match and best_start_pos != -1:
-                match_pos = best_start_pos
-        if not is_match:
-            return []
-
-    if not is_match:
-        return []
-
-    raw_hit_pos = _find_raw_hit_position(raw_for_positions, raw_keywords, space_mode)
-    return [_build_search_result(entry, raw_text, raw_for_positions, raw_hit_pos, include_detail)]
-
-
-def search_entries_chunk(
-    entries: List[Dict],
-    norm_keyword_groups: List[List[str]],
-    raw_keywords: List[str],
-    search_mode: str,
-    range_limit: int,
-    space_mode: str,
-    normalize_mode: str,
-    include_detail: bool,
-) -> List[Dict]:
-    chunk_results: List[Dict] = []
-    for entry in entries:
-        hit = search_text_logic(
-            entry,
-            norm_keyword_groups,
-            raw_keywords,
-            search_mode,
-            range_limit,
-            space_mode,
-            normalize_mode,
-            include_detail,
-        )
-        if hit:
-            chunk_results.extend(hit)
-    return chunk_results
-
-
-def build_query_groups(query: str, space_mode: str, normalize_mode: str) -> Tuple[List[str], List[List[str]]]:
-    keywords = [k for k in (query or "").split() if k.strip()]
-    normalizer = normalize_text_nfkc_casefold if normalize_mode == "normalized" else normalize_text_minimal
-    norm_keyword_groups = [
-        [apply_space_mode(normalizer(k).strip(), space_mode)]
-        for k in keywords
-        if k.strip()
-    ]
-    return keywords, norm_keyword_groups
-
-
-
-
-def check_range_match(norm_text: str, norm_keyword_groups: List[List[str]], range_limit: int) -> bool:
-    positions = []
-    for group_id, group in enumerate(norm_keyword_groups):
-        for kw in group:
-            if not kw:
-                continue
-            start = 0
-            while True:
-                idx = norm_text.find(kw, start)
-                if idx == -1:
-                    break
-                positions.append((idx, group_id))
-                start = idx + 1
-    positions.sort()
-    left = 0
-    count_map = defaultdict(int)
-    unique_count = 0
-    target_unique = len(norm_keyword_groups)
-    for right in range(len(positions)):
-        r_idx, r_group_id = positions[right]
-        if count_map[r_group_id] == 0:
-            unique_count += 1
-        count_map[r_group_id] += 1
-        while unique_count == target_unique:
-            l_idx, l_group_id = positions[left]
-            current_len = r_idx - l_idx
-            if current_len <= range_limit:
-                return True
-            count_map[l_group_id] -= 1
-            if count_map[l_group_id] == 0:
-                unique_count -= 1
-            left += 1
-    return False
-
-
-def search_text_logic_shared(
-    entry: Dict,
-    mm: mmap.mmap,
-    norm_keyword_groups: List[List[str]],
-    norm_keyword_groups_bytes: List[List[bytes]],
-    raw_keywords: List[str],
-    search_mode: str,
-    range_limit: int,
-    space_mode: str,
-    normalize_mode: str,
-    include_detail: bool,
-) -> List[Dict]:
-    if not norm_keyword_groups_bytes:
-        return []
-
-    if normalize_mode == "normalized":
-        if space_mode == "all":
-            norm_start = entry.get("norm_cf_all_off", entry["norm_all_off"])
-            norm_len = entry.get("norm_cf_all_len", entry["norm_all_len"])
-        elif space_mode == "jp":
-            norm_start = entry.get("norm_cf_jp_off", entry["norm_jp_off"])
-            norm_len = entry.get("norm_cf_jp_len", entry["norm_jp_len"])
-        else:
-            norm_start = entry.get("norm_cf_off", entry["norm_off"])
-            norm_len = entry.get("norm_cf_len", entry["norm_len"])
-    else:
-        if space_mode == "all":
-            norm_start = entry.get("norm_strict_all_off", entry["norm_all_off"])
-            norm_len = entry.get("norm_strict_all_len", entry["norm_all_len"])
-        elif space_mode == "jp":
-            norm_start = entry.get("norm_strict_jp_off", entry["norm_jp_off"])
-            norm_len = entry.get("norm_strict_jp_len", entry["norm_jp_len"])
-        else:
-            norm_start = entry.get("norm_strict_off", entry["norm_off"])
-            norm_len = entry.get("norm_strict_len", entry["norm_len"])
-    norm_end = norm_start + norm_len
-
-    is_match = False
-
-    if search_mode == "OR":
-        for group in norm_keyword_groups_bytes:
-            for k in group:
-                if not k:
-                    continue
-                if mm.find(k, norm_start, norm_end) != -1:
-                    is_match = True
-                    break
-            if is_match:
-                break
-    elif search_mode == "AND":
-        for group in norm_keyword_groups_bytes:
-            if not any(mm.find(k, norm_start, norm_end) != -1 for k in group if k):
-                return []
-        if range_limit == 0 or len(norm_keyword_groups_bytes) == 1:
-            is_match = True
-        else:
-            norm_text = mm[norm_start:norm_end].decode("utf-8", errors="ignore")
-            is_match = check_range_match(norm_text, norm_keyword_groups, range_limit)
-        if not is_match:
-            return []
-
-    if not is_match:
-        return []
-
-    raw_start = entry["raw_off"]
-    raw_len = entry["raw_len"]
-    raw_end = raw_start + raw_len
-    raw_bytes = mm[raw_start:raw_end]
-    raw_text = raw_bytes.decode("utf-8", errors="ignore")
-    raw_for_positions = normalize_invisible_separators(raw_text)
-
-    raw_hit_pos = _find_raw_hit_position(raw_for_positions, raw_keywords, space_mode)
-    return [_build_search_result(entry, raw_text, raw_for_positions, raw_hit_pos, include_detail)]
-
-
-def search_entries_chunk_shared(
-    entries: List[Dict],
-    mm: mmap.mmap,
-    norm_keyword_groups: List[List[str]],
-    norm_keyword_groups_bytes: List[List[bytes]],
-    raw_keywords: List[str],
-    search_mode: str,
-    range_limit: int,
-    space_mode: str,
-    normalize_mode: str,
-    include_detail: bool,
-) -> List[Dict]:
-    chunk_results: List[Dict] = []
-    for entry in entries:
-        hit = search_text_logic_shared(
-            entry,
-            mm,
-            norm_keyword_groups,
-            norm_keyword_groups_bytes,
-            raw_keywords,
-            search_mode,
-            range_limit,
-            space_mode,
-            normalize_mode,
-            include_detail,
-        )
-        if hit:
-            chunk_results.extend(hit)
-    return chunk_results
+from .extractors import extract_text_from_file_with_reason
+from .index_ops import (
+    build_index_for_folder,
+    cleanup_old_generations,
+    create_generation_uuid,
+    create_manifest,
+    get_current_generation_dir,
+    get_current_generation_manifest,
+    get_current_generation_pointer,
+    get_generation_dir,
+    index_path_for,
+    list_generations,
+    load_failures,
+    load_index_from_disk,
+    save_manifest,
+    scan_files,
+    set_current_generation_pointer,
+)
+from .search import (
+    build_query_groups,
+    init_search_worker,
+    init_search_worker_shared,
+    perform_search_process,
+    search_entries_chunk,
+)
+from .text_utils import (
+    apply_space_mode,
+    build_detail_text,
+    encode_norm_text,
+    normalize_invisible_separators,
+    normalize_snippet_text,
+    normalize_text,
+    normalize_text_minimal,
+    normalize_text_nfkc_casefold,
+)
+from .utils import (
+    colorize_url,
+    env_bool,
+    env_float,
+    env_int,
+    file_id_from_path,
+    folder_id_from_path,
+    get_ipv4_addresses,
+    log_info,
+    log_notice,
+    log_success,
+    log_warn,
+)
 
 
 def perform_search(
@@ -1003,208 +179,6 @@ def perform_search(
     return results
 
 
-def init_search_worker(folder_snapshot: List[Dict], aliases: Dict[str, str]):
-    global WORKER_MEMORY_PAGES, WORKER_FOLDER_META, host_aliases
-    host_aliases = aliases
-    pages_map: Dict[str, List[Dict]] = {}
-    meta_map: Dict[str, Tuple[str, str]] = {}
-    for folder in folder_snapshot:
-        folder_id = folder["id"]
-        folder_path = folder["path"]
-        folder_name = folder["name"]
-        cache = load_index_from_disk(folder_path)
-        pages_map[folder_id] = build_memory_pages(folder_id, folder_name, cache)
-        meta_map[folder_id] = (folder_path, folder_name)
-    WORKER_MEMORY_PAGES = pages_map
-    WORKER_FOLDER_META = meta_map
-
-
-def init_search_worker_shared(folder_snapshot: List[Dict], aliases: Dict[str, str]):
-    global WORKER_SHARED_ENTRIES, WORKER_SHARED_MM, WORKER_SHARED_FILES, host_aliases, PROCESS_SHARED
-    host_aliases = aliases
-    PROCESS_SHARED = True
-    WORKER_SHARED_ENTRIES = {}
-    WORKER_SHARED_MM = {}
-    WORKER_SHARED_FILES = {}
-    for folder in folder_snapshot:
-        folder_id = folder["id"]
-        shared_path = folder["shared_path"]
-        entries = folder["entries"]
-        if not entries:
-            WORKER_SHARED_ENTRIES[folder_id] = []
-            continue
-        try:
-            fh = open(shared_path, "rb")
-            mm = mmap.mmap(fh.fileno(), 0, access=mmap.ACCESS_READ)
-        except Exception:
-            log_info(f"共有ストア読込失敗: {shared_path}")
-            continue
-        WORKER_SHARED_FILES[folder_id] = fh
-        WORKER_SHARED_MM[folder_id] = mm
-        WORKER_SHARED_ENTRIES[folder_id] = entries
-
-
-def perform_search_process(
-    target_ids: List[str],
-    params: "SearchParams",
-    norm_keyword_groups: List[List[str]],
-    raw_keywords: List[str],
-    worker_count: int,
-    include_detail: bool,
-) -> List[Dict]:
-    if PROCESS_SHARED:
-        return perform_search_process_shared(
-            target_ids,
-            params,
-            norm_keyword_groups,
-            raw_keywords,
-            worker_count,
-            include_detail,
-        )
-
-    results: List[Dict] = []
-
-    def search_folder(fid: str) -> List[Dict]:
-        entries = WORKER_MEMORY_PAGES.get(fid)
-        if not entries:
-            meta = WORKER_FOLDER_META.get(fid)
-            if meta:
-                folder_path, folder_name = meta
-                cache = load_index_from_disk(folder_path)
-                entries = build_memory_pages(fid, folder_name, cache)
-                WORKER_MEMORY_PAGES[fid] = entries
-            else:
-                entries = []
-        folder_results: List[Dict] = []
-        entries_iter = entries
-        if len(entries_iter) >= SEARCH_ENTRIES_CHUNK_THRESHOLD:
-            chunk_workers = max(1, worker_count)
-            chunk_size = max(1, (len(entries_iter) + chunk_workers - 1) // chunk_workers)
-            chunks = [
-                entries_iter[i : i + chunk_size]
-                for i in range(0, len(entries_iter), chunk_size)
-            ]
-            with ThreadPoolExecutor(max_workers=chunk_workers) as executor:
-                futures = [
-                    executor.submit(
-                        search_entries_chunk,
-                        chunk,
-                        norm_keyword_groups,
-                        raw_keywords,
-                        params.mode,
-                        params.range_limit,
-                        params.space_mode,
-                        params.normalize_mode,
-                        include_detail,
-                    )
-                    for chunk in chunks
-                ]
-                for future in futures:
-                    try:
-                        res = future.result()
-                        if res:
-                            folder_results.extend(res)
-                    except Exception:
-                        continue
-        else:
-            folder_results = search_entries_chunk(
-                entries_iter,
-                norm_keyword_groups,
-                raw_keywords,
-                params.mode,
-                params.range_limit,
-                params.space_mode,
-                params.normalize_mode,
-                include_detail,
-            )
-        return folder_results
-
-    for fid in target_ids:
-        try:
-            res = search_folder(fid)
-            if res:
-                results.extend(res)
-        except Exception:
-            continue
-    return results
-
-
-def perform_search_process_shared(
-    target_ids: List[str],
-    params: "SearchParams",
-    norm_keyword_groups: List[List[str]],
-    raw_keywords: List[str],
-    worker_count: int,
-    include_detail: bool,
-) -> List[Dict]:
-    results: List[Dict] = []
-    norm_keyword_groups_bytes = [
-        [encode_norm_text(k) for k in group if k] for group in norm_keyword_groups
-    ]
-
-    def search_folder(fid: str) -> List[Dict]:
-        entries = WORKER_SHARED_ENTRIES.get(fid, [])
-        mm = WORKER_SHARED_MM.get(fid)
-        if not mm or not entries:
-            return []
-        folder_results: List[Dict] = []
-        entries_iter = entries
-        if len(entries_iter) >= SEARCH_ENTRIES_CHUNK_THRESHOLD:
-            chunk_workers = max(1, worker_count)
-            chunk_size = max(1, (len(entries_iter) + chunk_workers - 1) // chunk_workers)
-            chunks = [
-                entries_iter[i : i + chunk_size]
-                for i in range(0, len(entries_iter), chunk_size)
-            ]
-            with ThreadPoolExecutor(max_workers=chunk_workers) as executor:
-                futures = [
-                    executor.submit(
-                        search_entries_chunk_shared,
-                        chunk,
-                        mm,
-                        norm_keyword_groups,
-                        norm_keyword_groups_bytes,
-                        raw_keywords,
-                        params.mode,
-                        params.range_limit,
-                        params.space_mode,
-                        params.normalize_mode,
-                        include_detail,
-                    )
-                    for chunk in chunks
-                ]
-                for future in futures:
-                    try:
-                        res = future.result()
-                        if res:
-                            folder_results.extend(res)
-                    except Exception:
-                        continue
-        else:
-            folder_results = search_entries_chunk_shared(
-                entries_iter,
-                mm,
-                norm_keyword_groups,
-                norm_keyword_groups_bytes,
-                raw_keywords,
-                params.mode,
-                params.range_limit,
-                params.space_mode,
-                params.normalize_mode,
-                include_detail,
-            )
-        return folder_results
-
-    for fid in target_ids:
-        try:
-            res = search_folder(fid)
-            if res:
-                results.extend(res)
-        except Exception:
-            continue
-    return results
-
-
 def run_search_direct(
     query: str,
     params: "SearchParams",
@@ -1243,894 +217,6 @@ def run_search_direct(
 # --- インデックス関連 ---
 
 # --- 世代管理ユーティリティ ---
-def create_generation_uuid() -> str:
-    """新しい世代UUIDを生成（タイムスタンプ + ランダム）."""
-    import uuid
-    timestamp = int(time.time())
-    random_part = uuid.uuid4().hex[:8]
-    return f"{timestamp}_{random_part}"
-
-
-def get_current_generation_pointer() -> str | None:
-    """currentポインターファイルから世代名を取得."""
-    if not CURRENT_POINTER_FILE.exists():
-        return None
-    try:
-        gen_name = CURRENT_POINTER_FILE.read_text(encoding="utf-8").strip()
-        return gen_name if gen_name else None
-    except Exception:
-        return None
-
-
-def set_current_generation_pointer(gen_name: str) -> None:
-    """currentポインターファイルに世代名を設定（原子的）."""
-    temp_file = CURRENT_POINTER_FILE.with_suffix(".tmp")
-    try:
-        temp_file.write_text(gen_name, encoding="utf-8")
-        temp_file.replace(CURRENT_POINTER_FILE)
-    except Exception as e:
-        log_warn(f"currentポインター設定失敗: {e}")
-        if temp_file.exists():
-            temp_file.unlink()
-
-
-def get_generation_dir(gen_name: str | None = None, build: bool = False) -> Path:
-    """世代ディレクトリのパスを取得."""
-    if gen_name is None:
-        gen_name = get_current_generation_pointer()
-        if gen_name is None:
-            return INDEX_DIR
-    base = BUILD_DIR if build else INDEX_DIR
-    return base / f"gen_{gen_name}"
-
-
-def get_current_generation_dir() -> Path:
-    """現在の世代ディレクトリを取得（current.txt 欠落時は最新世代を自動選択）."""
-    gen_name = get_current_generation_pointer()
-    if gen_name:
-        gen_dir = get_generation_dir(gen_name, build=False)
-        if gen_dir.exists():
-            return gen_dir
-
-    # current.txt がない場合、最新の世代を自動選択
-    generations = list_generations()
-    if generations:
-        latest_gen_name, latest_gen_dir, latest_manifest = generations[0]  # 最新（ソート済み）
-        log_warn(f"current.txt が見つからないため、最新世代を使用: gen_{latest_gen_name}")
-        # current.txt を復旧
-        set_current_generation_pointer(latest_gen_name)
-        return latest_gen_dir
-
-    # 世代ディレクトリが一つもない場合のみ INDEX_DIR にフォールバック
-    return INDEX_DIR
-
-
-def create_manifest(gen_name: str, gen_dir: Path, folder_states_snapshot: Dict) -> Dict:
-    """manifest.json を作成."""
-    import datetime
-    now = datetime.datetime.now(datetime.timezone.utc)
-
-    folder_paths = sorted([meta["path"] for meta in folder_states_snapshot.values()])
-    source_folders_hash = hashlib.sha256(
-        "|".join(folder_paths).encode("utf-8")
-    ).hexdigest()
-
-    folders_info = {}
-    for folder_id, meta in folder_states_snapshot.items():
-        path = meta["path"]
-        hashed = hashlib.sha256(path.encode("utf-8")).hexdigest()[:12]
-        index_file = f"index_{hashed}_{INDEX_VERSION}.pkl.gz"
-        index_path = gen_dir / index_file
-        file_count = 0
-        if index_path.exists():
-            try:
-                with gzip.open(index_path, "rb") as f:
-                    data = pickle.load(f)
-                    if isinstance(data, dict):
-                        file_count = len(data)
-            except Exception:
-                pass
-
-        folders_info[folder_id] = {
-            "path": path,
-            "name": meta["name"],
-            "file_count": file_count,
-            "index_file": index_file,
-        }
-
-    manifest = {
-        "index_uuid": gen_name,
-        "schema_version": INDEX_VERSION,
-        "created_at": now.isoformat(),
-        "created_timestamp": int(now.timestamp()),
-        "source_folders_hash": source_folders_hash,
-        "folders": folders_info,
-    }
-
-    return manifest
-
-
-def save_manifest(gen_dir: Path, manifest: Dict) -> None:
-    """manifest.json を保存."""
-    manifest_path = gen_dir / "manifest.json"
-    try:
-        with manifest_path.open("w", encoding="utf-8") as f:
-            json.dump(manifest, f, ensure_ascii=False, indent=2)
-    except Exception as e:
-        log_warn(f"manifest.json 保存失敗: {e}")
-
-
-def load_manifest(gen_dir: Path) -> Dict | None:
-    """manifest.json を読み込み."""
-    manifest_path = gen_dir / "manifest.json"
-    if not manifest_path.exists():
-        return None
-    try:
-        with manifest_path.open("r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception:
-        return None
-
-
-def get_current_generation_manifest() -> Dict | None:
-    """現在の世代の manifest.json を取得."""
-    try:
-        gen_dir = get_current_generation_dir()
-        if not gen_dir or not gen_dir.exists():
-            return None
-        return load_manifest(gen_dir)
-    except Exception:
-        return None
-
-
-def list_generations() -> List[Tuple[str, Path, Dict | None]]:
-    """すべての世代をリストアップ（世代名, パス, manifest）."""
-    generations = []
-    for item in INDEX_DIR.iterdir():
-        if item.is_dir() and item.name.startswith("gen_"):
-            gen_name = item.name[4:]  # "gen_" を除く
-            manifest = load_manifest(item)
-            generations.append((gen_name, item, manifest))
-
-    # タイムスタンプでソート（新しい順）
-    # manifest がない場合は gen_name からタイムスタンプを抽出してフォールバック
-    def get_sort_key(item):
-        gen_name, gen_dir, manifest = item
-        if manifest and "created_timestamp" in manifest:
-            return manifest["created_timestamp"]
-        # manifest がない場合、gen_name からタイムスタンプを抽出
-        # gen_name format: "timestamp_randomhex"
-        try:
-            timestamp_str = gen_name.split("_")[0]
-            return int(timestamp_str)
-        except (ValueError, IndexError):
-            return 0
-
-    generations.sort(key=get_sort_key, reverse=True)
-    return generations
-
-
-def cleanup_old_generations(current_gen_name: str | None, grace_sec: int = 300) -> None:
-    """古い世代を削除（保持ポリシーに従う）."""
-    import shutil
-
-    keep_generations = env_int("INDEX_KEEP_GENERATIONS", 3)
-    keep_days = env_int("INDEX_KEEP_DAYS", 0)
-    max_bytes = env_int("INDEX_MAX_BYTES", 0)
-
-    all_generations = list_generations()
-
-    # 現在の世代を特定
-    current_gen_dir = None
-    if current_gen_name:
-        current_gen_dir = get_generation_dir(current_gen_name, build=False)
-
-    # 現在の世代を除外し、猶予期間も考慮したリストを作成
-    eligible_generations = []
-    for gen_name, gen_dir, manifest in all_generations:
-        # 現在の世代は絶対に削除しない
-        if current_gen_dir and gen_dir == current_gen_dir:
-            continue
-
-        # 猶予期間内は削除しない
-        if manifest and grace_sec > 0:
-            created_timestamp = manifest.get("created_timestamp", 0)
-            age_sec = time.time() - created_timestamp
-            if age_sec < grace_sec:
-                continue
-
-        eligible_generations.append((gen_name, gen_dir, manifest))
-
-    # 削除候補リスト
-    to_delete = []
-
-    # 世代数チェック（新しい順にソート済みなので、keep_generations より後ろを削除）
-    if keep_generations > 0 and len(eligible_generations) > keep_generations:
-        for gen_name, gen_dir, manifest in eligible_generations[keep_generations:]:
-            to_delete.append((gen_name, gen_dir, "世代数超過"))
-
-    # 日数チェック
-    if keep_days > 0:
-        for gen_name, gen_dir, manifest in eligible_generations:
-            if (gen_name, gen_dir, "世代数超過") in to_delete:
-                continue  # 既に削除対象
-            if manifest:
-                created_timestamp = manifest.get("created_timestamp", 0)
-                age_days = (time.time() - created_timestamp) / 86400
-                if age_days > keep_days:
-                    to_delete.append((gen_name, gen_dir, "保持期限切れ"))
-
-    # 容量チェック（現在の世代と猶予期間中の世代を含む全体のディスク使用量を計算）
-    if max_bytes > 0:
-        total_bytes = 0
-        keep_set = set((gen_name, gen_dir) for gen_name, gen_dir, _ in to_delete)
-
-        # 現在の世代のサイズを計算
-        if current_gen_dir and current_gen_dir.exists():
-            try:
-                current_bytes = sum(
-                    f.stat().st_size for f in current_gen_dir.rglob("*") if f.is_file()
-                )
-                total_bytes += current_bytes
-            except Exception:
-                pass
-
-        # すべての世代（猶予期間中も含む）のサイズを計算
-        # 削除対象に既に含まれているものは除外
-        for gen_name, gen_dir, manifest in all_generations:
-            # 現在の世代はスキップ（既に計算済み）
-            if current_gen_dir and gen_dir == current_gen_dir:
-                continue
-            # 削除対象のものはスキップ
-            if (gen_name, gen_dir) in keep_set:
-                continue
-            try:
-                dir_bytes = sum(
-                    f.stat().st_size for f in gen_dir.rglob("*") if f.is_file()
-                )
-                total_bytes += dir_bytes
-            except Exception:
-                pass
-
-        # 容量超過の場合、古い順に削除（猶予期間中と現在の世代は除外）
-        if total_bytes > max_bytes:
-            for gen_name, gen_dir, manifest in reversed(eligible_generations):
-                if (gen_name, gen_dir) in keep_set:
-                    continue  # 既に削除対象
-                try:
-                    dir_bytes = sum(
-                        f.stat().st_size for f in gen_dir.rglob("*") if f.is_file()
-                    )
-                    total_bytes -= dir_bytes
-                    to_delete.append((gen_name, gen_dir, "容量超過"))
-                    if total_bytes <= max_bytes:
-                        break
-                except Exception:
-                    pass
-
-    # 削除実行
-    for gen_name, gen_dir, reason in to_delete:
-        try:
-            shutil.rmtree(gen_dir)
-            log_info(f"世代削除: gen_{gen_name} 理由={reason}")
-        except Exception as e:
-            log_warn(f"世代削除失敗: gen_{gen_name} エラー={e}")
-
-
-def index_path_for(folder_path: str, gen_dir: Path | None = None) -> Path:
-    """インデックスファイルのパスを取得（世代ディレクトリ対応）."""
-    hashed = hashlib.sha256(folder_path.encode("utf-8")).hexdigest()[:12]
-    if gen_dir is None:
-        gen_dir = get_current_generation_dir()
-    return gen_dir / f"index_{hashed}_{INDEX_VERSION}.pkl.gz"
-
-
-def file_state_path_for(folder_path: str, gen_dir: Path | None = None) -> Path:
-    """file_state.jsonl のパスを取得（世代ディレクトリ対応）."""
-    hashed = hashlib.sha256(folder_path.encode("utf-8")).hexdigest()[:12]
-    if gen_dir is None:
-        gen_dir = get_current_generation_dir()
-    return gen_dir / f"file_state_{hashed}.jsonl"
-
-
-def failures_path_for(folder_path: str, gen_dir: Path | None = None) -> Path:
-    """failures.json のパスを取得（世代ディレクトリ対応）."""
-    hashed = hashlib.sha256(folder_path.encode("utf-8")).hexdigest()[:12]
-    if gen_dir is None:
-        gen_dir = get_current_generation_dir()
-    return gen_dir / f"failures_{hashed}.json"
-
-
-def load_file_state(folder_path: str, gen_dir: Path | None = None) -> Dict[str, Dict]:
-    """file_state.jsonl をディスクから読み込み（世代ディレクトリ対応）."""
-    state_path = file_state_path_for(folder_path, gen_dir)
-    if not state_path.exists():
-        return {}
-    states = {}
-    try:
-        with open(state_path, "r", encoding="utf-8") as f:
-            for line_num, line in enumerate(f, 1):
-                line = line.strip()
-                if not line:
-                    continue
-                try:
-                    entry = json.loads(line)
-                    # 必須キー "path" の確認
-                    if "path" not in entry:
-                        log_warn(f"file_state.jsonl 読み込み警告: 行{line_num} に 'path' キーがありません")
-                        continue
-                    states[entry["path"]] = entry
-                except json.JSONDecodeError:
-                    log_warn(f"file_state.jsonl 読み込み警告: 行{line_num} の JSON パースに失敗しました")
-                    continue
-                except Exception as e:
-                    log_warn(f"file_state.jsonl 読み込み警告: 行{line_num} の処理中にエラー ({e})")
-                    continue
-    except Exception as e:
-        log_warn(f"file_state.jsonl 読み込みエラー: {state_path} ({e})")
-    return states
-
-
-def save_file_state(folder_path: str, states: Dict[str, Dict], gen_dir: Path | None = None):
-    """file_state.jsonl をディスクに保存（世代ディレクトリ対応）."""
-    state_path = file_state_path_for(folder_path, gen_dir)
-    if gen_dir is not None:
-        gen_dir.mkdir(parents=True, exist_ok=True)
-    try:
-        with open(state_path, "w", encoding="utf-8") as f:
-            for path_str in sorted(states.keys()):
-                f.write(json.dumps(states[path_str], ensure_ascii=False) + "\n")
-    except Exception:
-        # 保存失敗時は静かにスキップ
-        return
-
-
-def load_failures(folder_path: str, gen_dir: Path | None = None) -> Tuple[Dict[str, str], bool]:
-    """failures.json をディスクから読み込み（世代ディレクトリ対応）."""
-    failures_path = failures_path_for(folder_path, gen_dir)
-    if not failures_path.exists():
-        return {}, False
-    try:
-        with open(failures_path, "r", encoding="utf-8") as f:
-            data = json.load(f)
-        if isinstance(data, dict):
-            return {str(k): str(v) for k, v in data.items()}, True
-    except Exception as e:
-        log_warn(f"failures_*.json 読み込みエラー: {failures_path} ({e})")
-    return {}, False
-
-
-def save_failures(folder_path: str, failures: Dict[str, str], gen_dir: Path | None = None):
-    """failures.json をディスクに保存（世代ディレクトリ対応）."""
-    failures_path = failures_path_for(folder_path, gen_dir)
-    if gen_dir is not None:
-        gen_dir.mkdir(parents=True, exist_ok=True)
-    temp_path = failures_path.with_suffix(".json.tmp")
-    try:
-        with open(temp_path, "w", encoding="utf-8") as f:
-            json.dump(failures, f, ensure_ascii=False)
-        os.replace(temp_path, failures_path)
-    except Exception as e:
-        log_warn(f"failures_*.json 保存失敗: {failures_path} ({e})")
-        try:
-            if temp_path.exists():
-                temp_path.unlink()
-        except Exception:
-            pass
-        return
-
-
-def get_file_stat(path_str: str) -> Dict:
-    """ファイルの stat 情報を取得（size, mtime_ns, inode/file_id）."""
-    try:
-        stat_info = os.stat(path_str)
-        result = {
-            "size": stat_info.st_size,
-            "mtime_ns": stat_info.st_mtime_ns,
-        }
-        # inode (Unix系) または file_id (Windows) を追加
-        if hasattr(stat_info, "st_ino") and stat_info.st_ino != 0:
-            result["inode"] = stat_info.st_ino
-        # Windowsの場合はfile indexを取得（可能なら）
-        if os.name == "nt":
-            try:
-                import ctypes
-                from ctypes import wintypes
-                kernel32 = ctypes.windll.kernel32
-                GENERIC_READ = 0x80000000
-                FILE_SHARE_READ = 0x00000001
-                OPEN_EXISTING = 3
-                FILE_FLAG_BACKUP_SEMANTICS = 0x02000000
-
-                handle = kernel32.CreateFileW(
-                    path_str,
-                    GENERIC_READ,
-                    FILE_SHARE_READ,
-                    None,
-                    OPEN_EXISTING,
-                    FILE_FLAG_BACKUP_SEMANTICS,
-                    None
-                )
-                if handle != -1:
-                    class BY_HANDLE_FILE_INFORMATION(ctypes.Structure):
-                        _fields_ = [
-                            ("dwFileAttributes", wintypes.DWORD),
-                            ("ftCreationTime", wintypes.FILETIME),
-                            ("ftLastAccessTime", wintypes.FILETIME),
-                            ("ftLastWriteTime", wintypes.FILETIME),
-                            ("dwVolumeSerialNumber", wintypes.DWORD),
-                            ("nFileSizeHigh", wintypes.DWORD),
-                            ("nFileSizeLow", wintypes.DWORD),
-                            ("nNumberOfLinks", wintypes.DWORD),
-                            ("nFileIndexHigh", wintypes.DWORD),
-                            ("nFileIndexLow", wintypes.DWORD),
-                        ]
-                    info = BY_HANDLE_FILE_INFORMATION()
-                    if kernel32.GetFileInformationByHandle(handle, ctypes.byref(info)):
-                        file_id = (info.nFileIndexHigh << 32) | info.nFileIndexLow
-                        result["file_id"] = file_id
-                    kernel32.CloseHandle(handle)
-            except Exception:
-                pass
-        return result
-    except Exception:
-        return {}
-
-
-def compute_fast_fingerprint(path_str: str, chunk_bytes: int = 65536) -> str | None:
-    """fast fingerprint を計算（先頭/末尾 N バイト + size）."""
-    try:
-        stat_info = os.stat(path_str)
-        file_size = stat_info.st_size
-
-        hasher = hashlib.sha256()
-        hasher.update(str(file_size).encode())
-
-        with open(path_str, "rb") as f:
-            # 先頭チャンク
-            head_chunk = f.read(chunk_bytes)
-            hasher.update(head_chunk)
-
-            # 末尾チャンク（ファイルが十分大きい場合）
-            if file_size > chunk_bytes:
-                f.seek(-min(chunk_bytes, file_size - len(head_chunk)), 2)
-                tail_chunk = f.read(chunk_bytes)
-                hasher.update(tail_chunk)
-
-        return hasher.hexdigest()
-    except Exception:
-        return None
-
-
-def compute_full_hash(path_str: str, algo: str = "sha256") -> str | None:
-    """full hash を計算."""
-    try:
-        hasher = hashlib.new(algo)
-        with open(path_str, "rb") as f:
-            while chunk := f.read(1048576):  # 1MB chunks
-                hasher.update(chunk)
-        return hasher.hexdigest()
-    except Exception:
-        return None
-
-
-def should_compute_full_hash(path_str: str) -> bool:
-    """full hash を計算すべきかどうか判定."""
-    # FULL_HASH_PATHS: カンマ区切りのパス接頭辞リスト
-    full_hash_paths = os.getenv("FULL_HASH_PATHS", "").strip()
-    if full_hash_paths:
-        for prefix in full_hash_paths.split(","):
-            prefix = prefix.strip()
-            if prefix and path_str.startswith(prefix):
-                return True
-
-    # FULL_HASH_EXTS: カンマ区切りの拡張子リスト
-    full_hash_exts = os.getenv("FULL_HASH_EXTS", "").strip()
-    if full_hash_exts:
-        ext = os.path.splitext(path_str)[1].lower()
-        for target_ext in full_hash_exts.split(","):
-            target_ext = target_ext.strip().lower()
-            if target_ext and ext == target_ext:
-                return True
-
-    # ネットワークドライブ判定（UNCパス or SMBマウント）
-    if os.name == "nt":
-        # Windows: UNCパス (\\server\share\...) または ネットワークドライブ
-        if path_str.startswith("\\\\") or path_str.startswith("//"):
-            return True
-    else:
-        # Unix系: /mnt, /net, /Volumes などの典型的なマウントポイント
-        mount_prefixes = ["/mnt/", "/net/", "/Volumes/"]
-        for prefix in mount_prefixes:
-            if path_str.startswith(prefix):
-                return True
-
-    return False
-
-
-def should_reindex_file(
-    path_str: str,
-    current_stat: Dict,
-    prev_state: Dict | None,
-    diff_mode: str,
-    fast_fp_bytes: int,
-    full_hash_algo: str,
-) -> Tuple[bool, Dict]:
-    """
-    段階的な差分判定を行い、再インデックスが必要かどうかを返す.
-
-    Returns:
-        (should_reindex, updated_state): 再インデックス要否と更新後の state
-    """
-    # 前回の状態がない場合は再インデックス必要
-    if not prev_state:
-        new_state = {
-            "path": path_str,
-            "size": current_stat.get("size"),
-            "mtime_ns": current_stat.get("mtime_ns"),
-            "last_indexed_at": time.time(),
-            "deletion_candidate_since": None,
-        }
-        if "inode" in current_stat:
-            new_state["inode"] = current_stat["inode"]
-        if "file_id" in current_stat:
-            new_state["file_id"] = current_stat["file_id"]
-        return True, new_state
-
-    # Stage 1: stat 比較
-    stat_changed = False
-    if current_stat.get("size") != prev_state.get("size"):
-        stat_changed = True
-    elif current_stat.get("mtime_ns") != prev_state.get("mtime_ns"):
-        stat_changed = True
-    elif "inode" in current_stat and "inode" in prev_state:
-        if current_stat["inode"] != prev_state["inode"]:
-            stat_changed = True
-    elif "file_id" in current_stat and "file_id" in prev_state:
-        if current_stat["file_id"] != prev_state["file_id"]:
-            stat_changed = True
-
-    if not stat_changed:
-        # stat が変わっていない場合は、削除候補をクリアして前回の状態を保持
-        updated_state = prev_state.copy()
-        updated_state["deletion_candidate_since"] = None
-        return False, updated_state
-
-    # stat モードの場合はここで終了（変更あり）
-    if diff_mode == "stat":
-        new_state = {
-            "path": path_str,
-            "size": current_stat.get("size"),
-            "mtime_ns": current_stat.get("mtime_ns"),
-            "last_indexed_at": time.time(),
-            "deletion_candidate_since": None,
-        }
-        if "inode" in current_stat:
-            new_state["inode"] = current_stat["inode"]
-        if "file_id" in current_stat:
-            new_state["file_id"] = current_stat["file_id"]
-        return True, new_state
-
-    # Stage 2: fast fingerprint 比較
-    if diff_mode in {"stat+fastfp", "stat+fastfp+fullhash"}:
-        current_fp = compute_fast_fingerprint(path_str, fast_fp_bytes)
-        prev_fp = prev_state.get("fast_fp")
-
-        if current_fp and prev_fp and current_fp == prev_fp:
-            # fast fingerprint が一致する場合は変更なし
-            updated_state = prev_state.copy()
-            updated_state["size"] = current_stat.get("size")
-            updated_state["mtime_ns"] = current_stat.get("mtime_ns")
-            if "inode" in current_stat:
-                updated_state["inode"] = current_stat["inode"]
-            if "file_id" in current_stat:
-                updated_state["file_id"] = current_stat["file_id"]
-            updated_state["deletion_candidate_since"] = None
-            return False, updated_state
-
-        # fast fingerprint が異なる、または計算できない場合
-        if diff_mode == "stat+fastfp":
-            new_state = {
-                "path": path_str,
-                "size": current_stat.get("size"),
-                "mtime_ns": current_stat.get("mtime_ns"),
-                "fast_fp": current_fp,
-                "last_indexed_at": time.time(),
-                "deletion_candidate_since": None,
-            }
-            if "inode" in current_stat:
-                new_state["inode"] = current_stat["inode"]
-            if "file_id" in current_stat:
-                new_state["file_id"] = current_stat["file_id"]
-            return True, new_state
-
-    # Stage 3: full hash 比較（条件付き）
-    if diff_mode == "stat+fastfp+fullhash" and should_compute_full_hash(path_str):
-        current_hash = compute_full_hash(path_str, full_hash_algo)
-        prev_hash = prev_state.get("full_hash")
-
-        if current_hash and prev_hash and current_hash == prev_hash:
-            # full hash が一致する場合は変更なし
-            updated_state = prev_state.copy()
-            updated_state["size"] = current_stat.get("size")
-            updated_state["mtime_ns"] = current_stat.get("mtime_ns")
-            if "inode" in current_stat:
-                updated_state["inode"] = current_stat["inode"]
-            if "file_id" in current_stat:
-                updated_state["file_id"] = current_stat["file_id"]
-            # fast fingerprint を更新
-            if diff_mode in {"stat+fastfp", "stat+fastfp+fullhash"}:
-                current_fp = compute_fast_fingerprint(path_str, fast_fp_bytes)
-                if current_fp:
-                    updated_state["fast_fp"] = current_fp
-            updated_state["deletion_candidate_since"] = None
-            return False, updated_state
-
-        # full hash が異なる、または計算できない場合
-        new_state = {
-            "path": path_str,
-            "size": current_stat.get("size"),
-            "mtime_ns": current_stat.get("mtime_ns"),
-            "last_indexed_at": time.time(),
-            "deletion_candidate_since": None,
-        }
-        if "inode" in current_stat:
-            new_state["inode"] = current_stat["inode"]
-        if "file_id" in current_stat:
-            new_state["file_id"] = current_stat["file_id"]
-        if diff_mode in {"stat+fastfp", "stat+fastfp+fullhash"}:
-            current_fp = compute_fast_fingerprint(path_str, fast_fp_bytes)
-            if current_fp:
-                new_state["fast_fp"] = current_fp
-        if current_hash:
-            new_state["full_hash"] = current_hash
-        return True, new_state
-
-    # それ以外の場合は変更ありとして処理
-    new_state = {
-        "path": path_str,
-        "size": current_stat.get("size"),
-        "mtime_ns": current_stat.get("mtime_ns"),
-        "last_indexed_at": time.time(),
-        "deletion_candidate_since": None,
-    }
-    if "inode" in current_stat:
-        new_state["inode"] = current_stat["inode"]
-    if "file_id" in current_stat:
-        new_state["file_id"] = current_stat["file_id"]
-    if diff_mode in {"stat+fastfp", "stat+fastfp+fullhash"}:
-        current_fp = compute_fast_fingerprint(path_str, fast_fp_bytes)
-        if current_fp:
-            new_state["fast_fp"] = current_fp
-    return True, new_state
-
-
-def load_index_from_disk(folder_path: str, gen_dir: Path | None = None) -> Dict[str, Dict]:
-    """インデックスをディスクから読み込み（世代ディレクトリ対応）."""
-    idx_path = index_path_for(folder_path, gen_dir)
-    if not idx_path.exists():
-        return {}
-    try:
-        with gzip.open(idx_path, "rb") as f:
-            data = pickle.load(f)
-            return data if isinstance(data, dict) else {}
-    except Exception:
-        return {}
-
-
-def save_index_to_disk(folder_path: str, cache: Dict[str, Dict], gen_dir: Path | None = None):
-    """インデックスをディスクに保存（世代ディレクトリ対応）."""
-    idx_path = index_path_for(folder_path, gen_dir)
-    # gen_dir が指定されている場合、ディレクトリを作成
-    if gen_dir is not None:
-        gen_dir.mkdir(parents=True, exist_ok=True)
-    try:
-        with gzip.open(idx_path, "wb") as f:
-            pickle.dump(cache, f)
-    except Exception:
-        # 保存失敗時は静かにスキップ
-        return
-
-
-def scan_files(folder: str) -> List[Path]:
-    root = Path(folder)
-    files = []
-    for f in root.rglob("*"):
-        if f.is_file() and f.suffix.lower() in ALLOWED_EXTS:
-            files.append(f)
-    return files
-
-
-def build_index_for_folder(folder: str, previous_failures: Dict[str, str] | None = None, gen_dir: Path | None = None, prev_gen_dir: Path | None = None, failures_loaded: bool | None = None) -> Tuple[Dict[str, Dict], Dict, Dict[str, str]]:
-    """インデックス作成（差分更新 & 高精度固定）."""
-    start_time = time.time()
-
-    # 段階的ハッシング設定を読み込み
-    diff_mode = os.getenv("DIFF_MODE", "stat").strip().lower()
-    if diff_mode not in {"stat", "stat+fastfp", "stat+fastfp+fullhash"}:
-        diff_mode = "stat"
-    try:
-        fast_fp_bytes = int(os.getenv("FAST_FP_BYTES", "65536").strip())
-        if fast_fp_bytes <= 0:
-            fast_fp_bytes = 65536
-    except (ValueError, AttributeError):
-        fast_fp_bytes = 65536
-    full_hash_algo = os.getenv("FULL_HASH_ALGO", "sha256").strip()
-    if not full_hash_algo:
-        full_hash_algo = "sha256"
-
-    # 差分更新のため、前回の世代から既存キャッシュとファイル状態を読み込む
-    if prev_gen_dir is not None and prev_gen_dir.exists():
-        existing_cache = load_index_from_disk(folder, prev_gen_dir)
-        prev_file_states = load_file_state(folder, prev_gen_dir)
-    else:
-        # フォールバック: 現在の世代から読み込む（初回起動時など）
-        existing_cache = load_index_from_disk(folder, gen_dir)
-        prev_file_states = load_file_state(folder, gen_dir)
-    failures = {k: v for k, v in (previous_failures or {}).items()}
-    all_files = scan_files(folder)
-    if not all_files and existing_cache:
-        stats = {
-            "total_files": len(existing_cache),
-            "indexed_files": len(existing_cache),
-            "updated_files": 0,
-            "skipped_files": len(existing_cache),
-        }
-        elapsed = time.time() - start_time
-        log_info(
-            f"インデックス構築スキップ: {folder} 件数={stats['indexed_files']} "
-            f"理由=ファイル一覧取得失敗 時間={elapsed:.1f}s"
-        )
-        failures = {k: v for k, v in failures.items() if k in existing_cache}
-        return existing_cache, stats, failures
-    if existing_cache and all_files:
-        allow_shrink = os.getenv("REBUILD_ALLOW_SHRINK", "").strip().lower() in {"1", "true", "yes"}
-        if not allow_shrink and len(all_files) < max(10, int(len(existing_cache) * 0.8)):
-            stats = {
-                "total_files": len(existing_cache),
-                "indexed_files": len(existing_cache),
-                "updated_files": 0,
-                "skipped_files": len(existing_cache),
-            }
-            elapsed = time.time() - start_time
-            log_info(
-                f"インデックス構築スキップ: {folder} 件数={stats['indexed_files']} "
-                f"理由=ファイル数急減 時間={elapsed:.1f}s"
-            )
-            failures = {k: v for k, v in failures.items() if k in existing_cache}
-            return existing_cache, stats, failures
-
-    current_map = {str(f): f for f in all_files}
-    failures = {k: v for k, v in failures.items() if k in current_map}
-    retry_set = set(failures.keys())
-    if prev_gen_dir is not None and failures_loaded is False:
-        bootstrap_set = {
-            path_str
-            for path_str in current_map
-            if path_str in prev_file_states and path_str not in existing_cache
-        }
-        log_notice(
-            "失敗履歴が無い/読み込めないため再試行対象を拡張: "
-            f"{folder} 件数={len(bootstrap_set)} "
-            f"(prev_states={len(prev_file_states)} cache={len(existing_cache)} files={len(current_map)})"
-        )
-        if bootstrap_set and not existing_cache and len(bootstrap_set) == len(current_map):
-            log_warn(f"再試行対象が全件になります: {folder}")
-        retry_set.update(bootstrap_set)
-
-    # valid_cache は既存のインデックスから開始（削除検知のため、まだフィルタしない）
-    valid_cache = dict(existing_cache)
-
-    # 削除検知の安全策: 削除候補の処理
-    current_file_states: Dict[str, Dict] = {}
-    for path_str in prev_file_states:
-        if path_str not in current_map:
-            # ファイルが見つからない
-            prev_state = prev_file_states[path_str]
-            if prev_state.get("deletion_candidate_since"):
-                # 2回目の不在 → 削除として扱う（インデックスから除外）
-                if path_str in valid_cache:
-                    del valid_cache[path_str]
-                    log_warn(f"削除確定（インデックス除外）: {path_str}")
-            else:
-                # 1回目の不在 → 削除候補としてマーク（インデックスには残す）
-                prev_state["deletion_candidate_since"] = time.time()
-                current_file_states[path_str] = prev_state
-                log_notice(f"削除候補: {path_str}")
-
-    # 段階的ハッシングによる差分判定
-    targets = []
-    target_set = set()
-
-    def add_target(path_str: str, path_obj: Path) -> None:
-        if path_str in target_set:
-            return
-        targets.append(path_obj)
-        target_set.add(path_str)
-
-    for path_str, path_obj in current_map.items():
-        current_stat = get_file_stat(path_str)
-        if not current_stat:
-            # stat 取得失敗 → 再インデックス対象
-            add_target(path_str, path_obj)
-            current_file_states[path_str] = {
-                "path": path_str,
-                "last_indexed_at": time.time(),
-                "deletion_candidate_since": None,
-            }
-            continue
-
-        prev_state = prev_file_states.get(path_str)
-        should_reindex, new_state = should_reindex_file(
-            path_str, current_stat, prev_state, diff_mode, fast_fp_bytes, full_hash_algo
-        )
-
-        if should_reindex or path_str in retry_set:
-            add_target(path_str, path_obj)
-
-        current_file_states[path_str] = new_state
-
-    updated_data: Dict[str, Dict] = {}
-    if targets:
-        max_workers = max(1, os.cpu_count() or 4)
-        log_info(f"インデックス更新開始: {folder} 対象={len(targets)} workers={max_workers}")
-        with ProcessPoolExecutor(max_workers=max_workers) as pool:
-            future_to_path = {
-                pool.submit(extract_text_from_file_with_reason, str(p)): str(p) for p in targets
-            }
-            for future in future_to_path:
-                path_str = future_to_path[future]
-                try:
-                    texts, reason = future.result()
-                    if texts:
-                        updated_data[path_str] = {
-                            "mtime": os.path.getmtime(path_str),
-                            "data": texts,
-                            "high_acc": True,
-                        }
-                        if path_str in failures:
-                            failures.pop(path_str, None)
-                        if path_str in retry_set:
-                            log_notice(f"再試行成功: {path_str}")
-                    else:
-                        failures[path_str] = reason or "抽出結果が空"
-                        log_warn(f"インデックス失敗: {path_str} 理由={failures[path_str]}")
-                except Exception:
-                    failures[path_str] = "例外"
-                    log_warn(f"インデックス失敗: {path_str} 理由=例外")
-                    continue
-        log_info(f"インデックス更新完了: {folder} 更新={len(updated_data)}")
-
-    valid_cache.update(updated_data)
-    failures = {k: v for k, v in failures.items() if k not in valid_cache}
-    save_index_to_disk(folder, valid_cache, gen_dir)
-
-    # ファイル状態を保存
-    save_file_state(folder, current_file_states, gen_dir)
-    save_failures(folder, failures, gen_dir)
-
-    stats = {
-        "total_files": len(all_files),
-        "indexed_files": len(valid_cache),
-        "updated_files": len(updated_data),
-        "skipped_files": len(valid_cache) - len(updated_data),
-    }
-    elapsed = time.time() - start_time
-    log_info(f"インデックス構築完了: {folder} 件数={stats['indexed_files']} 時間={elapsed:.1f}s")
-
-    # 失敗情報のログ出力（運用把握のため）
-    if failures:
-        # 詳細は SEARCH_DEBUG=1 の時のみ（大量ログ対策）
-        debug_mode = os.getenv("SEARCH_DEBUG", "").strip().lower() in {"1", "true", "yes"}
-        if debug_mode:
-            for i, (path, reason) in enumerate(list(failures.items())[:5], 1):
-                log_warn(f"  [{i}] {path}: {reason}")
-            if len(failures) > 5:
-                log_warn(f"  ... 他 {len(failures) - 5} 件")
-
-    return valid_cache, stats, failures
-
-
 def build_memory_pages(folder_id: str, folder_name: str, cache: Dict[str, Dict]) -> List[Dict]:
     store_normalized = env_bool("INDEX_STORE_NORMALIZED", True)
     pages: List[Dict] = []
@@ -2409,11 +495,9 @@ async def lifespan(app: FastAPI):
             f"検索実行モード: thread concurrency={concurrency} workers={workers} budget={budget}"
         )
     sync_state()
-    scheme = (
-        "https"
-        if (Path(os.getenv("CERT_DIR", "certs")) / "lan-cert.pem").exists()
-        else "http"
-    )
+    cert_dir = os.getenv("CERT_DIR", "certs")
+    cert_path = (BASE_DIR / cert_dir).resolve()
+    scheme = "https" if (cert_path / "lan-cert.pem").exists() else "http"
     urls = [
         f"{scheme}://{ip}:{os.getenv('PORT', '80')}"
         for ip in (get_ipv4_addresses() or ["0.0.0.0"])
@@ -2590,14 +674,6 @@ class AsyncRWLock:
             yield
         finally:
             self._writer_lock.release()
-
-
-def folder_id_from_path(path: str) -> str:
-    return hashlib.sha256(path.encode("utf-8")).hexdigest()[:10]
-
-
-def file_id_from_path(path: str) -> str:
-    return hashlib.sha256(path.encode("utf-8")).hexdigest()[:8]
 
 
 def register_file_id(path: str, folder_id: str) -> str:
@@ -2852,74 +928,6 @@ def install_asyncio_exception_filter():
             loop.default_exception_handler(context)
 
     loop.set_exception_handler(handler)
-
-
-def log_info(message: str):
-    if colorama_init:
-        colorama_init()
-    timestamp = time.strftime("%Y-%m-%d %H:%M:%S")
-    print(f"[{timestamp}] {message}")
-
-
-def log_warn(message: str):
-    if colorama_init:
-        colorama_init()
-    timestamp = time.strftime("%Y-%m-%d %H:%M:%S")
-    if Fore and Style:
-        print(f"{Fore.RED}[{timestamp}] {message}{Style.RESET_ALL}")
-    else:
-        print(f"[{timestamp}] {message}")
-
-
-def log_notice(message: str):
-    """黄色で表示する通知ログ（削除候補など、警告だが重要度は中程度）."""
-    if colorama_init:
-        colorama_init()
-    timestamp = time.strftime("%Y-%m-%d %H:%M:%S")
-    if Fore and Style:
-        print(f"{Fore.YELLOW}[{timestamp}] {message}{Style.RESET_ALL}")
-    else:
-        print(f"[{timestamp}] {message}")
-
-
-def log_success(message: str):
-    if colorama_init:
-        colorama_init()
-    timestamp = time.strftime("%Y-%m-%d %H:%M:%S")
-    if Fore and Style:
-        print(f"{Fore.GREEN}[{timestamp}] {message}{Style.RESET_ALL}")
-    else:
-        print(f"[{timestamp}] {message}")
-
-
-def colorize_url(url: str) -> str:
-    if Fore and Style:
-        return f"{Fore.CYAN}{url}{Style.RESET_ALL}"
-    return url
-
-
-def env_int(name: str, default: int) -> int:
-    raw = os.getenv(name, "").strip()
-    if raw.isdigit():
-        return int(raw)
-    return default
-
-
-def env_float(name: str, default: float) -> float:
-    raw = os.getenv(name, "").strip()
-    try:
-        return float(raw)
-    except Exception:
-        return default
-
-
-def env_bool(name: str, default: bool) -> bool:
-    raw = os.getenv(name, "").strip().lower()
-    if raw in {"1", "true", "yes"}:
-        return True
-    if raw in {"0", "false", "no"}:
-        return False
-    return default
 
 
 def resolve_normalize_mode(value: str | None) -> str:
@@ -3661,7 +1669,7 @@ async def get_folder_files(folder_id: str, scope: str = "indexed"):
 
 @app.get("/api/detail")
 async def get_detail(
-    file_id: str = Query(..., min_length=8, max_length=8, pattern=r"^[0-9a-f]{8}$"),
+    file_id: str = Query(..., min_length=12, max_length=12, pattern=r"^[0-9a-f]{12}$"),
     page: str = Query(..., min_length=1),
     hit_pos: int = Query(0, ge=0),
 ):
