+++ static/styles.css
@@ -1,7 +1,6 @@
 /* ═══════════════════════════════════════════════════════════════
    フォルダ内テキスト検索 — YomiToku Style Design System
-   System Version: 1.1.11
-   File Version: 1.1.6
+   System Version: 1.2.0
    ═══════════════════════════════════════════════════════════════ */
 
 :root {
